refactor: trash repository
