+++ server/src/services/trash.service.spec.ts
@@ -1,7 +1,7 @@
 import { BadRequestException } from '@nestjs/common';
 import { IJobRepository, JobName, JobStatus } from 'src/interfaces/job.interface';
-import { ITrashRepository } from 'src/interfaces/trash.interface';
 import { TrashService } from 'src/services/trash.service';
+import { ITrashRepository } from 'src/types';
 import { authStub } from 'test/fixtures/auth.stub';
 import { IAccessRepositoryMock } from 'test/repositories/access.repository.mock';
 import { newTestService } from 'test/utils';
